requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
anthropic==0.28.0
python-dotenv==1.0.1
psycopg2-binary
